requirements.txt
pcaspy==0.7.1
pyserial==3.4
PyYAML==3.12
